DEV: backend blogs improved activity 5.5-5.11

+++ blog/controllers/blogs.js
@@ -1,6 +1,15 @@
+const jwt = require('jsonwebtoken')
 const blogRouter = require('express').Router()
-const blog = require('../models/blog')
 const Blog = require('../models/blog')
+const User = require('../models/user')
+
+// const getTokenFrom = request => {
+//   const authorization = request.get('authorization')
+//   if (authorization && authorization.startsWith('Bearer ')) {
+//     return authorization.replace('Bearer ', '')
+//   }
+//   return null
+// }
 
 blogRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({})
@@ -19,32 +28,76 @@ blogRouter.get('/:id', async (request, response) =>{
 blogRouter.post('/', async (request, response) => {
   const body = request.body
 
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const user = request.user
+
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes || 0
+    likes: body.likes || 0,
+    user: user.id
   })
 
+  const userInfo = await User.findById(user.id)
   const savedBlog = await blog.save()
+  userInfo.blogs = userInfo.blogs.concat(savedBlog.id)
+  await userInfo.save() 
+
   response.status(201).json(savedBlog)
 })
 
 blogRouter.put('/:id', async (request, response) => {
     const body = request.body
 
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = request.user;
+
     const blog = {
       title: body.title,
       author: body.author,
       url: body.url,
-      likes: body.likes
+      likes: body.likes,
+      user: user.id
     }
     
+    const userInfo = await User.findById(user.id)
     await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
+    userInfo.blogs = userInfo.blogs.concat(blog.id)
+    await userInfo.save()
+    
     response.status(200).end()
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+
+  if (blog.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: 'unauthorized user' })
+  }
+  const userInfo = await User.findById(user.id);
+  
+  userInfo.blogs = userInfo.blogs.filter(b => b && b.toString() !== blog.id.toString())
+  await userInfo.save()
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
